+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/client/hook/ReceiveMsgHook.java
@@ -15,19 +15,15 @@
  * limitations under the License.
  */
 
-package client.common;
+package org.apache.eventmesh.runtime.client.hook;
 
+import io.netty.channel.ChannelHandlerContext;
 
-public interface ClientConstants {
+import org.apache.eventmesh.common.protocol.tcp.Package;
 
-    /**
-     * CLIENT HEART BEAT TIME
-     */
-    int HEARTBEAT = 1000 * 60;
-
-    long DEFAULT_TIMEOUT_IN_MILLISECONDS = 3000;
-
-    String SYNC_TOPIC = "TEST-TOPIC-TCP-SYNC";
-    String ASYNC_TOPIC = "TEST-TOPIC-TCP-ASYNC";
-    String BROADCAST_TOPIC = "TEST-TOPIC-TCP-BROADCAST";
+/**
+ * Business callback hook, which is a callback for all types of messages
+ */
+public interface ReceiveMsgHook {
+    void handle(Package msg, ChannelHandlerContext ctx);
 }
