+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/demo/SyncPubClient.java
@@ -20,9 +20,9 @@ package org.apache.eventmesh.runtime.demo;
 import org.apache.eventmesh.common.protocol.tcp.EventMeshMessage;
 import org.apache.eventmesh.common.protocol.tcp.Package;
 
-import client.common.MessageUtils;
-import client.common.UserAgentUtils;
-import client.impl.PubClientImpl;
+import org.apache.eventmesh.runtime.client.common.MessageUtils;
+import org.apache.eventmesh.runtime.client.common.UserAgentUtils;
+import org.apache.eventmesh.runtime.client.impl.PubClientImpl;
 
 public class SyncPubClient {
     public static void main(String[] args) throws Exception {
