+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/demo/SyncSubClient.java
@@ -23,10 +23,10 @@ import org.apache.eventmesh.common.protocol.SubcriptionType;
 import org.apache.eventmesh.common.protocol.tcp.Command;
 import org.apache.eventmesh.common.protocol.tcp.Package;
 
-import client.common.ClientConstants;
-import client.common.MessageUtils;
-import client.hook.ReceiveMsgHook;
-import client.impl.SubClientImpl;
+import org.apache.eventmesh.runtime.client.common.ClientConstants;
+import org.apache.eventmesh.runtime.client.common.MessageUtils;
+import org.apache.eventmesh.runtime.client.hook.ReceiveMsgHook;
+import org.apache.eventmesh.runtime.client.impl.SubClientImpl;
 import org.apache.eventmesh.common.protocol.SubscriptionMode;
 
 public class SyncSubClient {
